Add syscall interception basics
Also protects SIGSYS handlers from being called reentrantly.
This could be trimmed down by removing the NODEFER flag when
registering this signal handler, but then the process would
just crash with an unhandled SIGSYS if this happens. This way
a useful error message is displayed if the code within the
SIGSYS handler changes or the Rust runtime changes and starts
using a forbidden syscall.