Remove imperfect SIGSYS reentrancy handling
Even with the harness code removed by this commit, a SIGSYS
might be triggered during this function's return (e.g. if the
libc makes a syscall behind the curtains like sigreturn),
after the AtomicBool has been reset to "not running".

The ideal case would have been to detect such conditions and
display a useful error message, but it's better to crash with
an unhandled SIGSYS than going into infinite recursion of the
SIGSYS handler.