Add implementation of access() through OpenFile requests
For now, the access() syscall is implemented through one or more
OpenFile request(s) to the broker. This avoids adding more requests
and attack surface on the broker, but makes this syscall a bit more
costly. Any file descriptor obtained through OpenFile is close()d
without having to go through the broker, of course.